+++ tests/e2e/core-tests/specs/merchant/wp-admin-order-apply-coupon.test.js
@@ -56,7 +56,7 @@ const runOrderApplyCouponTest = () => {
 			// Verify the coupon list is showing
 			await page.waitForSelector('.wc-used-coupons');
 			await expect(page).toMatchElement('.wc_coupon_list', { text: 'Coupon(s)' });
-			await expect(page).toMatchElement('.wc_coupon_list li.code.editable', { text: couponCode });
+			await expect(page).toMatchElement('.wc_coupon_list li.code.editable', { text: couponCode.toLowerCase() });
 
 			// Check that the coupon has been applied
 			await expect(page).toMatchElement('.wc-order-item-discount', { text: '5.00' });
@@ -66,13 +66,13 @@ const runOrderApplyCouponTest = () => {
 		it('can remove a coupon', async () => {
 			// Make sure we have a coupon on the page to use
 			await page.waitForSelector('.wc-used-coupons');
-			await expect(page).toMatchElement('.wc_coupon_list li.code.editable', { text: couponCode });
+			await expect(page).toMatchElement('.wc_coupon_list li.code.editable', { text: couponCode.toLowerCase() });
 			await evalAndClick( 'a.remove-coupon' );
 
 			await uiUnblocked();
 
 			// Verify the coupon pricing has been removed
-			await expect(page).not.toMatchElement('.wc_coupon_list li.code.editable', { text: couponCode });
+			await expect(page).not.toMatchElement('.wc_coupon_list li.code.editable', { text: couponCode.toLowerCase() });
 			await expect(page).not.toMatchElement('.wc-order-item-discount', { text: '5.00' });
 			await expect(page).not.toMatchElement('.line-cost .view .woocommerce-Price-amount', { text: discountedPrice });